backend/requirements.txt
fastapi 
uvicorn
transformers
torch
sentencepiece
python-dotenv


